+++ coroutines-run-blocking-all/jvm/main/kotlin/com/javiersc/kotlinx/coroutines/run/blocking/RunBlocking.kt
@@ -1,0 +1,11 @@
+package com.javiersc.kotlinx.coroutines.run.blocking
+
+import kotlin.coroutines.CoroutineContext
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.runBlocking
+
+/** JVM `runBlocking` implementation */
+public actual fun <T> runBlocking(
+    context: CoroutineContext,
+    block: suspend CoroutineScope.() -> T
+): T = runBlocking(context, block)