+++ coroutines-run-blocking/common/test/kotlin/com/javiersc/kotlinx/coroutines/run/blocking/RunBlockingTest.kt
@@ -1,0 +1,12 @@
+package com.javiersc.kotlinx.coroutines.run.blocking
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+import kotlinx.coroutines.yield
+
+internal class RunBlockingTest {
+
+    @Test fun testRunBlocking() = runBlocking { assertTrue { getBoolean() } }
+}
+
+internal suspend fun getBoolean(): Boolean = yield().run { true }